change type of OSM id

+++ vwsfriend/vwsfriend/model/charging_session.py
@@ -1,6 +1,6 @@
 import enum
 
-from sqlalchemy import Column, Integer, Float, String, DateTime, Enum, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, Float, String, DateTime, Enum, ForeignKey
 from sqlalchemy.orm import relationship
 
 from vwsfriend.model.base import Base
@@ -38,7 +38,7 @@ class ChargingSession(Base):
     mileage_km = Column(Integer)
     position_latitude = Column(Float)
     position_longitude = Column(Float)
-    location_id = Column(Integer, ForeignKey('locations.osm_id'))
+    location_id = Column(BigInteger, ForeignKey('locations.osm_id'))
     location = relationship("Location")
 
     def __init__(self, vehicle):
